+++ src/headless/converse-core.js
@@ -11,7 +11,6 @@ import Promise from "es6-promise/dist/es6-promise.auto";
 import _ from "./lodash.noconflict";
 import advancedFormat from 'dayjs/plugin/advancedFormat'
 import dayjs from "dayjs";
-import f from "./lodash.fp";
 import i18n from "./i18n";
 import pluggable from "pluggable.js/src/pluggable";
 import polyfill from "./polyfill";
@@ -548,7 +547,7 @@ function cleanup () {
 _converse.initialize = async function (settings, callback) {
     settings = !_.isUndefined(settings) ? settings : {};
     const init_promise = u.getResolveablePromise();
-    _.each(PROMISES, addPromise);
+    PROMISES.forEach(addPromise);
     if (!_.isUndefined(_converse.connection)) {
         cleanup();
     }
@@ -724,18 +723,40 @@ _converse.initialize = async function (settings, callback) {
         _converse.api.send(pres);
     };
 
-    this.reconnect = _.debounce(function () {
-        _converse.log('RECONNECTING');
-        _converse.log('The connection has dropped, attempting to reconnect.');
+
+    /**
+     * Called once the XMPP connection has dropped and we want to attempt
+     * reconnection.
+     * @method reconnect
+     * @private
+     * @memberOf _converse
+     */
+    this.reconnect = _.debounce(() => {
+        _converse.log('RECONNECTING: the connection has dropped, attempting to reconnect.');
         _converse.setConnectionStatus(
             Strophe.Status.RECONNECTING,
             __('The connection has dropped, attempting to reconnect.')
         );
+        /**
+         * Triggered when the connection has dropped, but Converse will attempt
+         * to reconnect again.
+         *
+         * @event _converse#will-reconnect
+         */
+        _converse.api.trigger('will-reconnect');
+
         _converse.connection.reconnecting = true;
         _converse.tearDown();
         _converse.logIn(null, true);
     }, 2000, {'leading': true});
 
+
+    /**
+     * Properly tear down the session so that it's possible to manually connect again.
+     * @method finishDisconnection
+     * @private
+     * @memberOf _converse
+     */
     this.finishDisconnection = function () {
         _converse.log('DISCONNECTED');
         delete _converse.connection.reconnecting;
@@ -751,19 +772,22 @@ _converse.initialize = async function (settings, callback) {
         _converse.api.trigger('disconnected');
     };
 
-    this.onDisconnected = function () {
-        /* Gets called once strophe's status reaches Strophe.Status.DISCONNECTED.
-         * Will either start a teardown process for converse.js or attempt
-         * to reconnect.
-         */
-        const reason = _converse.disconnection_reason;
 
+    /**
+     * Gets called once strophe's status reaches Strophe.Status.DISCONNECTED.
+     * Will either start a teardown process for converse.js or attempt
+     * to reconnect.
+     * @method onDisconnected
+     * @private
+     * @memberOf _converse
+     */
+    this.onDisconnected = function () {
+        const reason = _converse.disconnection_reason;
         if (_converse.disconnection_cause === Strophe.Status.AUTHFAIL) {
             if (_converse.credentials_url && _converse.auto_reconnect) {
                 /* In this case, we reconnect, because we might be receiving
                  * expirable tokens from the credentials_url.
                  */
-                _converse.api.trigger('will-reconnect');
                 return _converse.reconnect();
             } else {
                 return _converse.finishDisconnection();
@@ -775,15 +799,9 @@ _converse.initialize = async function (settings, callback) {
                 !_converse.auto_reconnect) {
             return _converse.finishDisconnection();
         }
-        /**
-         * Triggered when the connection has dropped, but Converse will attempt
-         * to reconnect again.
-         *
-         * @event _converse#will-reconnect
-         */
-        _converse.api.trigger('will-reconnect');
         _converse.reconnect();
     };
+
 
     this.setDisconnectionCause = function (cause, reason, override) {
         /* Used to keep track of why we got disconnected, so that we can
@@ -922,7 +940,7 @@ _converse.initialize = async function (settings, callback) {
             _converse.tearDown();
         }
         // Recreate all the promises
-        _.each(Object.keys(_converse.promises), addPromise);
+        Object.keys(_converse.promises).forEach(addPromise);
         /**
          * Triggered once the user has logged out.
          * @event _converse#logout
@@ -1655,8 +1673,8 @@ _converse.api = {
          * @example _converse.api.promises.add('foo-completed');
          */
         'add' (promises) {
-            promises = _.isArray(promises) ? promises : [promises]
-            _.each(promises, addPromise);
+            promises = Array.isArray(promises) ? promises : [promises]
+            promises.forEach(addPromise);
         }
     },
 
@@ -1908,7 +1926,6 @@ const converse = {
      * @property {function} converse.env.Strophe   - The [Strophe](http://strophe.im/strophejs) XMPP library used by Converse.
      * @property {object} converse.env._           - The instance of [lodash](http://lodash.com) used by Converse.
      * @property {function} converse.env.f         - And instance of Lodash with its methods wrapped to produce immutable auto-curried iteratee-first data-last methods.
-     * @property {function} converse.env.b64_sha1  - Utility method from Strophe for creating base64 encoded sha1 hashes.
      * @property {object} converse.env.dayjs       - [DayJS](https://github.com/iamkun/dayjs) date manipulation library.
      * @property {function} converse.env.sizzle    - [Sizzle](https://sizzlejs.com) CSS selector engine.
      * @property {object} converse.env.utils       - Module containing common utility methods used by Converse.
@@ -1922,7 +1939,6 @@ const converse = {
         'Promise': Promise,
         'Strophe': Strophe,
         '_': _,
-        'f': f,
         'b64_sha1':  b64_sha1,
         'dayjs': dayjs,
         'sizzle': sizzle,